requirements.txt
attrs==19.3.0
Automat==0.8.0
autopep8==1.6.0
black==21.12b0
blinker==1.4
certifi==2019.11.28
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cloud-init==21.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==36.0.1
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
flake8==4.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.2.0
mypy-extensions==0.4.3
netifaces==0.10.4
oauthlib==3.1.0
pathspec==0.9.0
pexpect==4.6.0
platformdirs==2.4.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.0.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
PyYAML==5.3.1
requests==2.27.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.1
ssh-import-id==5.10
systemd-python==234
toml==0.10.2
tomli==1.2.3
Twisted==18.9.0
typing-extensions==4.0.1
ubuntu-advantage-tools==27.2
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
